# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Production ready starter guide.",
        "title": "My Go API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/session-pauses": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "session_pauses"
                ],
                "summary": "Create a new session pause",
                "parameters": [
                    {
                        "description": "Session pause info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateSessionPauseRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.SessionPauseResponse"
                        }
                    }
                }
            }
        },
        "/session-pauses/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "session_pauses"
                ],
                "summary": "Get a session pause by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pause ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.SessionPauseResponse"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "session_pauses"
                ],
                "summary": "Delete a session pause",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pause ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/session-pauses/{id}/end": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "session_pauses"
                ],
                "summary": "End a session pause",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pause ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "End pause info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.EndSessionPauseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/study-cycles": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "study_cycles"
                ],
                "summary": "Create a new study cycle",
                "parameters": [
                    {
                        "description": "Study cycle info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateStudyCycleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.StudyCycleResponse"
                        }
                    }
                }
            }
        },
...
        "/study-cycles/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "study_cycles"
                ],
                "summary": "Get a study cycle by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cycle ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.StudyCycleResponse"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "study_cycles"
                ],
                "summary": "Update a study cycle",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cycle ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Study cycle info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateStudyCycleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "study_cycles"
                ],
                "summary": "Delete a study cycle",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cycle ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.CreateSessionPauseRequest": {
            "type": "object",
            "required": [
                "session_id",
                "started_at"
            ],
            "properties": {
                "session_id": {
                    "type": "string"
                },
                "started_at": {
                    "type": "string"
                }
            }
        },
        "handler.CreateStudyCycleRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string",
                    "minLength": 2
                }
            }
        },
...
        "handler.EndSessionPauseRequest": {
            "type": "object",
            "required": [
                "ended_at"
            ],
            "properties": {
                "ended_at": {
                    "type": "string"
                }
            }
        },
...
        "handler.SessionPauseResponse": {
            "type": "object",
            "properties": {
                "ended_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "session_id": {
                    "type": "string"
                },
                "started_at": {
                    "type": "string"
                }
            }
        },
        "handler.StudyCycleResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "handler.UpdateStudyCycleRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string",
                    "minLength": 2
                }
            }
        },
...
    }
}